# Fix path traversal bypass