--- a/library/helpers/safeDecodeURIComponent.ts
+++ b/library/helpers/safeDecodeURIComponent.ts
@@ -1,0 +1,7 @@
+export function safeDecodeURIComponent(uri: string): string | undefined {
+  try {
+    return decodeURIComponent(uri);
+  } catch {
+    return undefined;
+  }
+}