Benchmark results:
- Before: `0.011ms`
- After, without url encoded strings: `0.012ms`
- After, with multiple valid url encoded strings: `0.012ms`
- After, with multiple valid and multiple invalid url encoded strings: `0.012ms`

`safeDecodeURIComponent` is around 94% faster than `decodeURIComponent`